refactor app into modules learn import/export
