+++ rusty-clock/output/timelog_2025-09-22.json
@@ -1,0 +1,10 @@
+[
+  [
+    "2025-09-22T08:39:35.045315+02:00",
+    "2025-09-22T08:39:51.695989+02:00"
+  ],
+  [
+    "2025-09-22T08:39:53.556213+02:00",
+    "2025-09-22T08:39:59.547521+02:00"
+  ]
+]
