+++ rusty-clock/output/timelog_2025-09-23.json
@@ -1,0 +1,6 @@
+[
+  [
+    "2025-09-23T12:37:01.965203+02:00",
+    "2025-09-23T12:37:14.158376+02:00"
+  ]
+]
