+++ rusty-clock/output/timelog_2025-09-27.json
@@ -1,0 +1,10 @@
+[
+  [
+    "2025-09-27T18:38:10.937546+02:00",
+    "2025-09-27T18:38:12.919741+02:00"
+  ],
+  [
+    "2025-09-27T18:38:14.014881+02:00",
+    "2025-09-27T18:38:17.819436+02:00"
+  ]
+]
