+++ rusty-clock/src/model.rs
@@ -1,0 +1,15 @@
+use chrono::{DateTime, Local};
+
+pub struct RustyClock {
+    pub start_time: Option<DateTime<Local>>,
+    pub log: Vec<(DateTime<Local>, DateTime<Local>)>,
+}
+
+impl Default for RustyClock {
+    fn default() -> Self {
+        Self {
+            start_time: None,
+            log: vec![],
+        }
+    }
+}
